requirements.txt
openai
python-dotenv
PyMuPDF
pillow
